requirements.txt
Flask==3.0.0
Jinja2==3.1.3
MarkupSafe==2.1.3
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-flask==0.43b0
opentelemetry-instrumentation-requests==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
protobuf==4.25.2
requests==2.31.0
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1